pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proxy-inference-engine"
version = "0.1.0"
description = "Optimized LLM & MLX inference for Apple Silicon."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "The Proxy Company", email = "info@theproxy.co"}
]
keywords = ["mlx", "inference", "llm", "apple", "vision", "language", "model", "ai", "proxy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [
    "pse",
    "mlx",
    "pillow",
    "requests",
    "transformers",
    "huggingface-hub",
    "fastapi[standard]",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "build",
    "twine",
]

[project.urls]
"Homepage" = "https://github.com/TheProxyCompany/proxy-inference-engine"
"Bug Tracker" = "https://github.com/TheProxyCompany/proxy-inference-engine/issues"
"Documentation" = "https://github.com/TheProxyCompany/proxy-inference-engine"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
